requirements.txt
h2o-wave>=0.24.0
requests>=2.31.0
python-dotenv>=1.0.0
pytest>=7.4.0
pytest-cov>=4.1.0 